@@ -1,6 +1,6 @@
 {
   "name": "@aliceo2/bookkeeping",
-  "version": "0.73.0",
+  "version": "0.74.0",
   "author": "ALICEO2",
   "scripts": {
     "coverage": "nyc npm test && npm run coverage:report",
@@ -40,7 +40,7 @@
   "devDependencies": {
     "@types/d3": "7.4.0",
     "chai": "4.3.4",
-    "eslint": "8.54.0",
+    "eslint": "8.55.0",
     "js-yaml": "4.1.0",
     "mocha": "10.2.0",
     "nodemon": "3.0.1",
