@@ -11,20 +11,12 @@
  *  or submit itself to any jurisdiction.
  */
 
-const { NotFoundError } = require('../../errors/NotFoundError.js');
-const { getLog } = require('./getLog.js');
+import { h } from '/js/src/index.js';
 
 /**
- * Fetch a log by its id and throws an error if none is found
- * @param {number} logId the log id
- * @return {Promise<SequelizeLog>} the log found
+ * Returns a component centered in its parent
+ *
+ * @param {Component} children the children of the centered component
+ * @return {Component} the resulting component
  */
-exports.getLogOrFail = async (logId) => {
-    const logModel = await getLog(logId);
-
-    if (logModel !== null) {
-        return logModel;
-    } else {
-        throw new NotFoundError(`Log with this id (${logId}) could not be found`);
-    }
-};
+export const centeredComponent = (children) => h('.w-100.h-100.flex-row.items-center.justify-center', children);
