@@ -21,17 +21,39 @@ const { timestampToMysql } = require('../../server/utilities/timestampToMysql.js
  * @param {Period} the period in which logs should have been created
  * @return {string} the query
  */
-const getTagDistributionQuery = ({ from, to }) => `
+const getTagDistributionQuery = ({ from, to }) => sequelize.query(`
     SELECT t.text tag, COUNT(*) count
     FROM logs l
              INNER JOIN log_tags lt on l.id = lt.log_id
              INNER JOIN tags t on lt.tag_id = t.id
-    WHERE l.created_at >= '${timestampToMysql(from)}'
-      AND l.created_at < '${timestampToMysql(to)}'
+    WHERE l.created_at >= :fromTimeStamp
+      AND l.created_at < :toTimeStamp
       AND t.archived_at IS NULL
     GROUP BY t.text
     ORDER BY COUNT(*) DESC, t.text
-`;
+`, {
+    replacements: { fromTimeStamp: timestampToMysql(from), toTimeStamp: timestampToMysql(to) },
+});
+
+/**
+ * Returns the query to fetch the reply count for a given array of logIds.
+ *
+ * @param {Array<number>} logIds The log ids for which the reply count will be fetched.
+ * @returns {Promise<Array<{ originLogId: number, replyCounter: number }>>} A promise resolving to an array of objects,
+ * each containing for each originLogId the reply count.
+ */
+const getLogsWithReplyCountQuery = (logIds) => sequelize.query(`
+    WITH RECURSIVE log_tree (originLogId, id) AS (
+        SELECT id originId, id FROM logs
+        WHERE id IN (:logIds)
+        UNION ALL
+        SELECT originLogId, l.id FROM logs as l
+            JOIN log_tree AS lt ON lt.id = l.parent_log_id
+    )
+    SELECT originLogId, COUNT(*) - 1 as replyCount FROM log_tree GROUP BY originLogId;
+`, {
+    replacements: { logIds },
+});
 
 /**
  * Sequelize implementation of the LogRepository.
@@ -45,21 +67,24 @@ class LogRepository extends Repository {
     }
 
     /**
-     * Adds a count of root log by their amount of children onto every applicable log.
+     * Adds a 'replies' property to the given rows array which denotes the amount of children for each applicable log.
      *
-     * @param {Array} rows The collection of rows already fetched
-     * @param {QueryBuilder} queryBuilder The QueryBuilder to use.
-     * @returns {Promise} Promise object representing the full mock data
+     * @param {Array} rows The collection of log rows already fetched.
+     * @returns {Promise} Promise returning logs with a 'replies' property, denoting the reply count for each log.
      */
-    async addChildrenCountByRootLog(rows, queryBuilder = new QueryBuilder()) {
-        const rootLogIds = [...new Set(rows.map((row) => row.rootLogId))];
-        queryBuilder
-            .where('rootLogId').oneOf(...rootLogIds)
-            .set('group', ['rootLogId']);
+    async addChildrenCountByLog(rows) {
+        const logIds = [...new Set(rows.map((row) => row.id))];
 
-        const replies = await this.count(queryBuilder);
-        replies.forEach(({ rootLogId, count }) => {
-            const rowToAttachTo = rows.find(({ id }) => id === rootLogId);
+        if (logIds.length === 0) {
+            return rows;
+        }
+
+        const [results] = await getLogsWithReplyCountQuery(logIds);
+
+        results.forEach((result) => {
+            const count = result.replyCount || 0;
+            const rowToAttachTo = rows.find(({ id }) => id === result.originLogId);
+
             if (rowToAttachTo) {
                 rowToAttachTo.replies = count;
             }
@@ -109,7 +134,7 @@ class LogRepository extends Repository {
      * @return {Promise<{tag: string, count: number}[]>} the tag and their amount
      */
     async getTagDistribution(period) {
-        const [rows] = await sequelize.query(getTagDistributionQuery(period));
+        const [rows] = await getTagDistributionQuery(period);
         return rows;
     }
 }
