@@ -15,7 +15,7 @@ import { Loader, Observable, QueryRouter, sessionService } from '/js/src/index.j
 
 import Logs from './views/Logs/Logs.js';
 import Tags from './views/Tags/Tags.js';
-import Runs from './views/Runs/Runs.js';
+import { RunsModel } from './views/Runs/RunsModel.js';
 import Subsystems from './views/Subsystems/Subsystems.js';
 import Flps from './views/Flps/Flps.js';
 import { EnvironmentModel } from './views/Environments/EnvironmentModel.js';
@@ -90,10 +90,7 @@ export default class Model extends Observable {
         this.lhcFills = new LhcFills(this);
         this.lhcFills.bubbleTo(this);
 
-        /**
-         * @type {RunModel}
-         */
-        this.runs = new Runs(this);
+        this.runs = new RunsModel(this);
         this.runs.bubbleTo(this);
 
         this.subsystems = new Subsystems(this);
@@ -158,7 +155,7 @@ export default class Model extends Observable {
             case 'home':
                 // Setting the rows per page also collects all the runs and logs
                 this.logs.fetchAllLogs();
-                this.runs.fetchAllRuns();
+                this.runs.loadOverview();
                 this.lhcFills.loadStableOverview();
                 break;
             case 'lhc-period-overview':
@@ -180,7 +177,6 @@ export default class Model extends Observable {
                 // Prevent loading extra rows when it's not necessary
                 if (!this.logs.pagination.isInfiniteScrollEnabled) {
                     this.logs.fetchAllLogs();
-                    this.runs.fetchAllRuns();
                 }
                 break;
             case 'log-detail':
@@ -190,10 +186,7 @@ export default class Model extends Observable {
                 this.logs.loadCreation(this.router.params);
                 break;
             case 'run-overview':
-                // Prevent loading extra rows when it's not necessary
-                if (!this.runs.pagination.isInfiniteScrollEnabled) {
-                    this.runs.fetchAllRuns();
-                }
+                this.runs.loadOverview();
                 break;
             case 'run-detail':
                 this.runs.loadDetails(this.router.params);
@@ -282,18 +275,6 @@ export default class Model extends Observable {
      */
     isAdmin() {
         return this.session.access.includes('admin');
-    }
-
-    /**
-     * Display a modal on the screen
-     * @see ModalModel
-     *
-     * @param {Modal} modal the modal to display
-     *
-     * @returns {void}
-     */
-    modal(modal) {
-        this.modalModel.display(modal);
     }
 
     /**
