@@ -21,15 +21,15 @@ today.setMinutes(today.getMinutes() - today.getTimezoneOffset());
 
 /**
  * Returns the creation date filter components
- * @param {RunModel} runModel the run model object
+ * @param {RunsOverviewModel} runsOverviewModel the run model object
  * @return {vnode} Two date selection boxes to control the minimum and maximum creation dates for the log filters
  */
-const o2endFilter = (runModel) => {
+const o2endFilter = (runsOverviewModel) => {
     const date = new Date();
-    const o2from = runModel.getO2endFilterFrom();
-    const o2to = runModel.getO2endFilterTo();
-    const o2fromTime = runModel.getO2endFilterFromTime();
-    const o2toTime = runModel.getO2endFilterToTime();
+    const o2from = runsOverviewModel.getO2endFilterFrom();
+    const o2to = runsOverviewModel.getO2endFilterTo();
+    const o2fromTime = runsOverviewModel.getO2endFilterFromTime();
+    const o2toTime = runsOverviewModel.getO2endFilterToTime();
     const now = `${date.getHours()}:${(date.getMinutes() < 10 ? '0' : '') + date.getMinutes()}`;
     return h('', [
         h('.f6', 'Ended from:'),
@@ -39,7 +39,7 @@ const o2endFilter = (runModel) => {
             placeholder: DATE_FORMAT,
             max: o2to || today,
             value: o2from,
-            oninput: (e) => runModel.setO2endFilter('From', e.target.value, e.target.validity.valid),
+            oninput: (e) => runsOverviewModel.setO2endFilter('From', e.target.value, e.target.validity.valid),
         }, ''),
         h('input.w-50.mv1', {
             type: 'time',
@@ -51,10 +51,10 @@ const o2endFilter = (runModel) => {
             oninput: (e) => {
                 const time = e.target.value ? e.target.value : '00:00';
                 if (!o2from) {
-                    runModel.setO2endFilter('From', today, true);
-                    runModel.setO2endFilter('FromTime', time, e.target.value <= now);
+                    runsOverviewModel.setO2endFilter('From', today, true);
+                    runsOverviewModel.setO2endFilter('FromTime', time, e.target.value <= now);
                 } else {
-                    runModel.setO2endFilter('FromTime', time, e.target.validity.valid);
+                    runsOverviewModel.setO2endFilter('FromTime', time, e.target.validity.valid);
                 }
             },
         }, ''),
@@ -65,7 +65,7 @@ const o2endFilter = (runModel) => {
             min: o2from,
             max: today,
             value: o2to,
-            oninput: (e) => runModel.setO2endFilter('To', e.target.value, e.target.validity.valid),
+            oninput: (e) => runsOverviewModel.setO2endFilter('To', e.target.value, e.target.validity.valid),
         }, ''),
 
         h('input.w-50.mv1', {
@@ -78,9 +78,9 @@ const o2endFilter = (runModel) => {
                 const time = e.target.value ? e.target.value :
                     today == o2from ? now : '23:59';
                 if (!o2to) {
-                    runModel.setO2endFilter('To', today, true);
+                    runsOverviewModel.setO2endFilter('To', today, true);
                 }
-                runModel.setO2endFilter('ToTime', time, e.target.validity.valid);
+                runsOverviewModel.setO2endFilter('ToTime', time, e.target.validity.valid);
             },
         }, ''),
     ]);
