@@ -88,8 +88,13 @@ class LineChartClassComponent {
 
         const { left, top } = this._svgNode.getBoundingClientRect();
         const relativeCursor = { x: event.clientX - left, y: event.clientY - top };
-        const { coordinates, relativePosition } = this._chartRenderer.getHoveredPoint(relativeCursor) || {};
-        this.updateHoveredPoint(coordinates, relativePosition ? { x: relativePosition.x + left, y: relativePosition.y + top } : null);
+
+        if (!this._svgNode.parentNode.contains(event.target)) {
+            this.updateHoveredPoint();
+        } else {
+            const { coordinates, relativePosition } = this._chartRenderer.getHoveredPoint(relativeCursor) || {};
+            this.updateHoveredPoint(coordinates, relativePosition ? { x: relativePosition.x + left, y: relativePosition.y + top } : null);
+        }
     }
 
     /**
@@ -149,12 +154,12 @@ class LineChartClassComponent {
     /**
      * Update the currently hovered value
      *
-     * @param {Point} coordinates the domain coordinates of the point being hovered
-     * @param {Point} absolutePosition the absolute coordinates (relative to the page) of the point being hovered
+     * @param {Point} [coordinates] the domain coordinates of the point being hovered
+     * @param {Point} [absolutePosition] the absolute coordinates (relative to the page) of the point being hovered
      * @return {void}
      */
     updateHoveredPoint(coordinates, absolutePosition) {
-        if (this.hoveredPointCoordinates !== coordinates) {
+        if (JSON.stringify(this.hoveredPointCoordinates) !== JSON.stringify(coordinates)) {
             if (coordinates === undefined) {
                 this.hoveredPointAbsolutePosition = null;
             } else {
