@@ -15,6 +15,7 @@ import { quadtree } from '/assets/d3.min.js';
 import { ChartRenderer } from './ChartRenderer.js';
 import { renderDatasetAsLine } from './dataset/renderDatasetAsLine.js';
 import { renderDatasetAsPointCloud } from './dataset/renderDatasetAsPointCloud.js';
+import { renderDatasetLegend } from './legend/renderDatasetLegend.js';
 
 /**
  * @typedef LineChartDatasetConfiguration
@@ -68,7 +69,7 @@ export class LineChartRenderer extends ChartRenderer {
             getY: this._stackedY ? ({ y }) => this.yScale(y[datasetIndex]) : ({ y }) => this.yScale(y),
         };
 
-        const { line: lineConfiguration, point: pointConfiguration } = this._datasets[datasetIndex] || {};
+        const { line: lineConfiguration, point: pointConfiguration, legend: legendConfiguration } = this._datasets[datasetIndex] || {};
 
         // Display datasets
 
@@ -79,8 +80,6 @@ export class LineChartRenderer extends ChartRenderer {
             renderDatasetAsLine(
                 svg,
                 points.filter((point) => (pointLocator.getY(point) ?? null) !== null),
-                this.xScale,
-                this.yScale,
                 pointLocator,
                 this._chartDrawingZone,
                 lineConfiguration,
@@ -90,6 +89,11 @@ export class LineChartRenderer extends ChartRenderer {
         // Display point cloud
         if (pointConfiguration) {
             renderDatasetAsPointCloud(svg, points, pointLocator, this._chartDrawingZone, pointConfiguration);
+        }
+
+        // Display dataset legend
+        if (legendConfiguration) {
+            renderDatasetLegend(svg, points, pointLocator, this._chartDrawingZone, legendConfiguration);
         }
 
         const datasetQuadtree = quadtree()
