@@ -25,8 +25,6 @@ import { getUniqueId } from '../../../../../utilities/getUniqueId.js';
  *
  * @param {SVGElement} svg the svg in which graph must be rendered
  * @param {Point[]} points the dataset to display
- * @param {d3.AxisScale} xScale the horizontal scale of the dataset
- * @param {d3.AxisScale} yScale the vertical scale of the dataset
  * @param {PointLocator} pointLocator locator to compute the points coordinates
  * @param {BoundingBox} chartDrawingZone the chart drawing zone outside which points are not rendered
  * @param {LineGraphConfiguration} [configuration] the line graph configuration
@@ -35,8 +33,6 @@ import { getUniqueId } from '../../../../../utilities/getUniqueId.js';
 export const renderDatasetAsLine = (
     svg,
     points,
-    xScale,
-    yScale,
     pointLocator,
     chartDrawingZone,
     configuration = {},
@@ -59,17 +55,22 @@ export const renderDatasetAsLine = (
         .x(pointLocator.getX)
         .y(pointLocator.getY)(points);
 
-    const xLeft = chartDrawingZone.left;
-    const xRight = chartDrawingZone.left + chartDrawingZone.width;
-    const yBottom = chartDrawingZone.top + chartDrawingZone.height;
-    if (points.length > 1) {
-        pathDraw = `M ${xLeft},${yBottom} ${pathDraw.substring(1)} ${xRight},${yBottom}`;
+    if (configuration.fill) {
+        const xLeft = chartDrawingZone.left;
+        const xRight = chartDrawingZone.left + chartDrawingZone.width;
+        const yBottom = chartDrawingZone.top + chartDrawingZone.height;
+        if (points.length > 1) {
+            pathDraw = `M ${xLeft},${yBottom} ${pathDraw.substring(1)} ${xRight},${yBottom}`;
+        }
     }
 
     path.attr('d', pathDraw);
 
     if (configuration.stroke) {
         path.attr('stroke', configuration.stroke);
+    }
+    if (configuration.thickness) {
+        path.attr('stroke-width', configuration.thickness);
     }
     if (configuration.fill) {
         path.attr('fill', configuration.fill);
