@@ -34,7 +34,7 @@ export const formatTimeRange = ({ from, to }, configuration) => {
     let parts = [];
 
     if (from === undefined && to === undefined) {
-        parts = '-';
+        parts = ['-'];
     } else if (from === undefined) {
         parts = [formatText('Before'), formatTimestamp(to)];
     } else if (to === undefined) {
