@@ -12,30 +12,34 @@
  */
 import { Observable, RemoteData } from '/js/src/index.js';
 import { getRemoteData } from './getRemoteData.js';
+import { ObservableData } from '../ObservableData.js';
 
 /**
  * Service class providing observable data fetching process
  */
-export class RemoteDataFetcher extends Observable {
+export class RemoteDataSource extends Observable {
     /**
      * Constructor
      */
     constructor() {
         super();
-        this._data = RemoteData.notAsked();
+
+        /**
+         * @type {ObservableData<RemoteData>}
+         * @private
+         */
+        this._observableData = new ObservableData(RemoteData.notAsked());
         this._abortController = null;
     }
 
     /**
      * Fetch the given endpoint to fill current data
      *
-     * @template T
      * @param {string} endpoint the endpoint to fetch
-     * @return {Promise<T>} resolves once the data fetching has ended
+     * @return {Promise<void>} resolves once the data fetching has ended
      */
     async fetch(endpoint) {
-        this._data = RemoteData.loading();
-        this.notify();
+        this._observableData.current = RemoteData.loading();
 
         const abortController = new AbortController();
         try {
@@ -44,25 +48,22 @@ export class RemoteDataFetcher extends Observable {
             }
             this._abortController = abortController;
             const { data } = await getRemoteData(endpoint, { signal: this._abortController.signal });
-            this._data = RemoteData.success(data);
-            this.notify();
-            return data;
+            this._observableData.current = RemoteData.success(data);
         } catch (error) {
             // Use local variable because the class member (this._abortController) may already have been override in another call
             if (!abortController.signal.aborted) {
-                this._data = RemoteData.failure(error);
+                this._observableData.current = RemoteData.failure(error);
             }
-            this.notify();
             throw error;
         }
     }
 
     /**
-     * Return the current data
+     * Return the observable remote data provided by this source
      *
-     * @return {RemoteData} the current remote data fetched by the fetcher
+     * @return {ObservableData<RemoteData>} the observable remote data
      */
-    get data() {
-        return this._data;
+    get observableData() {
+        return this._observableData;
     }
 }
