@@ -13,6 +13,7 @@
 
 import { formatTimestamp } from '../../../utilities/formatting/formatTimestamp.js';
 import { formatRunsList } from '../../Runs/format/formatRunsList.js';
+import { displayEnvironmentStatusHistory } from '../format/displayEnvironmentStatusHistory.js';
 import { displayEnvironmentStatus } from '../format/displayEnvironmentStatus.js';
 import { displayEnvironmentInfologgerLinks } from '../format/displayEnvironmentInfologgerLinks.js';
 import { frontLink } from '../../../components/common/navigation/frontLink.js';
@@ -46,20 +47,20 @@ export const environmentsActiveColumns = {
         format: (timestamp) => formatTimestamp(timestamp),
     },
     status: {
-        name: 'Status',
+        name: 'Current Status',
         visible: true,
         sortable: false,
-        size: 'w-10',
+        size: 'w-20',
         noEllipsis: true,
         format: (_, environment) => displayEnvironmentStatus(environment),
     },
-    statusMessage: {
-        name: 'Status Message',
+    statusHistory: {
+        name: 'Status History',
         visible: true,
         sortable: false,
-        size: 'w-40',
+        size: 'w-30',
         title: true,
-        format: (message) => message || '-',
+        format: (_, environment) => displayEnvironmentStatusHistory(environment),
         balloon: true,
     },
     runs: {
