@@ -16,11 +16,12 @@ import { h } from '/js/src/index.js';
 /**
  * Returns a component displaying a colored environment status
  *
- * @param {string} status The environment status to display
+ * @param {string} status The environment status color to display
+ * @param {string} content Optional custom content to show in given status color
  *
  * @return {vnode} the formatted status
  */
-export const coloredEnvironmentStatusComponent = (status) => h(statusToCssClassMapping[status] ?? '', status);
+export const coloredEnvironmentStatusComponent = (status, content = status) => h(statusToCssClassMapping[status] ?? '', content);
 
 const statusToCssClassMapping = {
     RUNNING: '.success',
