@@ -27,16 +27,18 @@ const PAGE_USED_HEIGHT = 155;
  * @param {Object} model Pass the model to access the defined functions
  * @return {vnode} Return the view of the page with the filtering options
  */
-const HomeOverviewScreen = (model) => {
+const HomeOverviewScreen = ({ logs: logsModel, runs: { overviewModel: runsOverviewModel },
+    lhcFills: { overviewModel: LhcFillsOverviewModel } }) => {
     const MAX_ROWS = 10;
     const rowCount = estimateDisplayableRowsCount(TABLEROW_HEIGHT, PAGE_USED_HEIGHT);
-    model.logs.pagination.provideDefaultItemsPerPage(rowCount, MAX_ROWS);
-    model.runs.pagination.provideDefaultItemsPerPage(rowCount);
-    model.lhcFills.overviewModel.pagination.provideDefaultItemsPerPage(rowCount);
 
-    const logs = model.logs.getLogs();
-    const runs = model.runs.getRuns();
-    const fills = model.lhcFills.overviewModel.lhcFills;
+    logsModel.pagination.provideDefaultItemsPerPage(rowCount, MAX_ROWS);
+    runsOverviewModel.pagination.provideDefaultItemsPerPage(rowCount);
+    LhcFillsOverviewModel.pagination.provideDefaultItemsPerPage(rowCount);
+
+    const logs = logsModel.getLogs();
+    const { runs } = runsOverviewModel;
+    const fills = LhcFillsOverviewModel.lhcFills;
 
     return (
         h('div', [
@@ -49,7 +51,7 @@ const HomeOverviewScreen = (model) => {
                 ]),
                 h('.flex-column.w-50', [
                     h('h3', 'Runs'),
-                    table(runs, runsActiveColumns(model.runs), null, { profile: 'home' }),
+                    table(runs, runsActiveColumns, null, { profile: 'home' }),
                 ]),
             ]),
             h('.flex-row.w-100.mv3', [
@@ -57,7 +59,7 @@ const HomeOverviewScreen = (model) => {
                     style: 'padding-right: 1rem;',
                 }, [
                     h('h3', 'LHC Fills'),
-                    table(fills, lhcFillsActiveColumns(model.lhcFills), null),
+                    table(fills, lhcFillsActiveColumns, null),
                 ]),
             ]),
         ]));
