@@ -18,6 +18,7 @@ import { frontLink } from '../../../components/common/navigation/frontLink.js';
 import { formatPercentage } from '../../../utilities/formatting/formatPercentage.js';
 import { h } from '/js/src/index.js';
 import { formatRunsList } from '../../Runs/format/formatRunsList.js';
+import { formatLhcFillsTimeLoss } from '../format/formatLhcFillsTimeLoss.js';
 
 /**
  * List of active columns for a lhc fills table
@@ -31,19 +32,19 @@ export const lhcFillsActiveColumns = {
         format: (fillNumber) => frontLink(fillNumber, 'lhc-fill-details', { fillNumber: fillNumber }),
     },
     stableBeamsStart: {
-        name: 'Stable beams start',
+        name: 'SB START',
         visible: true,
         size: 'w-8',
-        format: (timestamp) => formatTimestamp(timestamp),
+        format: (timestamp) => formatTimestamp(timestamp, false),
     },
     stableBeamsEnd: {
-        name: 'Stable beams end',
+        name: 'SB END',
         visible: true,
         size: 'w-8',
-        format: (timestamp) => formatTimestamp(timestamp),
+        format: (timestamp) => formatTimestamp(timestamp, false),
     },
     stableBeamsDuration: {
-        name: 'Beams Duration',
+        name: 'SB Duration',
         visible: true,
         size: 'w-8',
         format: (time, { stableBeamsStart, stableBeamsEnd }) => {
@@ -74,13 +75,13 @@ export const lhcFillsActiveColumns = {
         name: 'Before 1st run',
         visible: true,
         size: 'w-8',
-        format: (duration, lhcFill) => duration ? `${formatDuration(duration)} (${formatPercentage(lhcFill.efficiencyLossAtStart)})` : '-',
+        format: (duration, lhcFill) => formatLhcFillsTimeLoss(duration, lhcFill.efficiencyLossAtStart, false),
     },
     timeLossAtEnd: {
         name: 'After last run',
         visible: true,
         size: 'w-8',
-        format: (duration, lhcFill) => duration ? `${formatDuration(duration)} (${formatPercentage(lhcFill.efficiencyLossAtEnd)})` : '-',
+        format: (duration, lhcFill) => formatLhcFillsTimeLoss(duration, lhcFill.efficiencyLossAtEnd, false),
     },
     meanRunDuration: {
         name: 'Mean run duration',
@@ -99,6 +100,7 @@ export const lhcFillsActiveColumns = {
         visible: true,
         size: 'w-10',
         format: (value) => value ? value : '-',
+        balloon: true,
     },
     runs: {
         name: 'Runs',
