@@ -12,22 +12,22 @@
  */
 
 import { Observable, RemoteData } from '/js/src/index.js';
-import { createCSVExport, createJSONExport } from '../../utilities/export.js';
-import { TagFilterModel } from '../../components/Filters/common/TagFilterModel.js';
-import { debounce } from '../../utilities/debounce.js';
-import { PaginationModel } from '../../components/Pagination/PaginationModel.js';
-import { getRemoteDataSlice } from '../../utilities/fetch/getRemoteDataSlice.js';
-import { DetectorsFilterModel } from '../../components/Filters/RunsFilter/DetectorsFilterModel.js';
-import { RunDetailsModel } from './Details/RunDetailsModel.js';
-import { RunTypesSelectionDropdownModel } from '../../components/runTypes/RunTypesSelectionDropdownModel.js';
-import { EorReasonFilterModel } from '../../components/Filters/RunsFilter/EorReasonsFilterModel.js';
+import { createCSVExport, createJSONExport } from '../../../utilities/export.js';
+import { TagFilterModel } from '../../../components/Filters/common/TagFilterModel.js';
+import { debounce } from '../../../utilities/debounce.js';
+import { PaginationModel } from '../../../components/Pagination/PaginationModel.js';
+import { getRemoteDataSlice } from '../../../utilities/fetch/getRemoteDataSlice.js';
+import { DetectorsFilterModel } from '../../../components/Filters/RunsFilter/DetectorsFilterModel.js';
+import { RunTypesSelectionDropdownModel } from '../../../components/runTypes/RunTypesSelectionDropdownModel.js';
+import { EorReasonFilterModel } from '../../../components/Filters/RunsFilter/EorReasonsFilterModel.js';
+import pick from '../../../utilities/pick.js';
 
 /**
  * Model representing handlers for runs page
  *
  * @implements {OverviewModel}
  */
-export default class RunModel extends Observable {
+export class RunsOverviewModel extends Observable {
     /**
      * The constructor of the Overview model object
      * @param {Model} model Pass the model to access the defined functions
@@ -35,13 +35,6 @@ export default class RunModel extends Observable {
      */
     constructor(model) {
         super();
-        this.model = model;
-
-        /**
-         * @type {RunDetailsModel}
-         * @private
-         */
-        this._detailsModel = null;
 
         this._listingTagsFilterModel = new TagFilterModel();
         this._listingTagsFilterModel.observe(() => this._applyFilters(true));
@@ -59,7 +52,7 @@ export default class RunModel extends Observable {
         this._eorReasonsFilterModel.observe(() => this._applyFilters());
 
         this._pagination = new PaginationModel();
-        this._pagination.observe(() => this.fetchAllRuns());
+        this._pagination.observe(() => this.fetchRuns());
         this._pagination.itemsPerPageSelector$.observe(() => this.notify());
 
         // Content
@@ -69,41 +62,17 @@ export default class RunModel extends Observable {
         this.reset(false);
 
         // eslint-disable-next-line no-return-assign,require-jsdoc
-        const updateDebounceTime = () => this._debouncedFetchAllRuns = debounce(this.fetchAllRuns.bind(this), model.inputDebounceTime);
+        const updateDebounceTime = () => this._debouncedFetchAllRuns = debounce(this.fetchRuns.bind(this), model.inputDebounceTime);
         model.appConfiguration$.observe(() => updateDebounceTime());
         updateDebounceTime();
     }
 
     /**
-     * Load the details page for the given run
-     * @param {string} [params.runNumber = null]  the runNumber of the run to display
-     * @param {string} [params.id = null]  Usage of this paramter is deprecated, use runNumber if feasible,
-     * @param {string} [params.panel = null] the key of the panel to display
-     * @return {void}
-     */
-    loadDetails({ runNumber = null, id: runId = null, panel: panelKey = null }) {
-        runId = runId ? Number(runId) : runId;
-        runNumber = runNumber ? Number(runNumber) : runNumber;
-
-        if (!this._detailsModel) {
-            this._detailsModel = new RunDetailsModel();
-            this._detailsModel.bubbleTo(this);
-        }
-        this._detailsModel.clearAndLoad({ runId, runNumber, panelKey });
-    }
-
-    /**
      * Retrieve every relevant run from the API
-     *
-     * @param {boolean} clear if true, any previous data will be discarded, even in infinite mode
-     *
-     * @return {undefined} Injects the data object with the response data
+     * @return {void} Injects the data object with the response data
      */
-    async fetchAllRuns(clear = false) {
-        /**
-         * @type {Run[]}
-         */
-        const concatenateWith = !clear && this._pagination.currentPage !== 1 && this._pagination.isInfiniteScrollEnabled
+    async fetchRuns() {
+        const concatenateWith = this._pagination.currentPage !== 1 && this._pagination.isInfiniteScrollEnabled
             ? this._currentPageRuns.payload || []
             : [];
 
@@ -111,8 +80,6 @@ export default class RunModel extends Observable {
             this._currentPageRuns = RemoteData.loading();
             this.notify();
         }
-
-        this._allRuns = RemoteData.notAsked();
 
         const params = {
             ...this._getFilterQueryParams(),
@@ -134,15 +101,25 @@ export default class RunModel extends Observable {
 
     /**
      * Create the export with the variables set in the model, handling errors appropriately
-     * @param {Object} content The source content.
+     * @param {object[]} runs The source content.
      * @param {string} fileName The name of the file including the output format.
+     * @param {Object<string, Function<*, string>>} exportFormats defines how particual fields of data units will be formated
      * @return {void}
      */
-    async createRunsExport(content, fileName) {
-        if (content.length > 0) {
+    async createRunsExport(runs, fileName, exportFormats) {
+        if (runs.length > 0) {
+            const selectedRunsFields = this.getSelectedRunsFields() || [];
+            runs = runs.map((selectedRun) => {
+                const entries = Object.entries(pick(selectedRun, selectedRunsFields));
+                const formattedEntries = entries.map(([key, value]) => {
+                    const formatExport = exportFormats[key].exportFormat || ((identity) => identity);
+                    return [key, formatExport(value, selectedRun)];
+                });
+                return Object.fromEntries(formattedEntries);
+            }),
             this.getSelectedExportType() === 'CSV'
-                ? createCSVExport(content, `${fileName}.csv`, 'text/csv;charset=utf-8;')
-                : createJSONExport(content, `${fileName}.json`, 'application/json');
+                ? createCSVExport(runs, `${fileName}.csv`, 'text/csv;charset=utf-8;')
+                : createJSONExport(runs, `${fileName}.json`, 'application/json');
         } else {
             this._currentPageRuns = RemoteData.failure([
                 {
@@ -152,14 +129,6 @@ export default class RunModel extends Observable {
             ]);
             this.notify();
         }
-    }
-
-    /**
-     * Getter for all the run data
-     * @return {RemoteData} Returns all of the filtered runs
-     */
-    getRuns() {
-        return this.runs;
     }
 
     /**
@@ -845,7 +814,7 @@ export default class RunModel extends Observable {
      */
     get allRuns() {
         if (this._allRuns.isNotAsked()) {
-            this._fetchAllRunsWithoutPaging();
+            this._fetchAllRunsWithoutPagination();
         }
 
         return this._allRuns;
@@ -898,15 +867,6 @@ export default class RunModel extends Observable {
      */
     get pagination() {
         return this._pagination;
-    }
-
-    /**
-     * Returns the run details sub models
-     *
-     * @return {RunDetailsModel|null} the sub-model
-     */
-    get detailsModel() {
-        return this._detailsModel;
     }
 
     /**
@@ -1005,7 +965,7 @@ export default class RunModel extends Observable {
      * @return {Promise<void>} void
      * @private
      */
-    async _fetchAllRunsWithoutPaging() {
+    async _fetchAllRunsWithoutPagination() {
         if (this.runs.isSuccess() && this.runs.payload.length === this._pagination.itemsCount) {
             this._allRuns = RemoteData.success([...this.runs.payload]);
             this.notify();
@@ -1037,6 +997,6 @@ export default class RunModel extends Observable {
      */
     _applyFilters(now = false) {
         this._pagination.currentPage = 1;
-        now ? this.fetchAllRuns() : this._debouncedFetchAllRuns(true);
+        now ? this.fetchRuns() : this._debouncedFetchAllRuns(true);
     }
 }
