@@ -24,34 +24,26 @@ const TABLEROW_HEIGHT = 59;
 const PAGE_USED_HEIGHT = 215;
 
 /**
- * Table row header
- * @param {Model} model Pass the model to access the defined functions
- * @return {vnode} Return the view of the table with the filtering options
- */
-const runsOverview = (model) => h('', showRunsTable(model, model.runs.getRuns()));
-
-/**
  * Build components in case of runs retrieval success
  * @param {Model} model model to access global functions
- * @param {RemoteData<Run[]>} runs list of runs retrieved from server
- * @return {vnode[]} Returns a vnode with the table containing the runs
+ * @return {Component} Returns a vnode with the table containing the runs
  */
-const showRunsTable = (model, runs) => {
-    model.runs.pagination.provideDefaultItemsPerPage(estimateDisplayableRowsCount(
+export const RunsOverviewPage = ({ runs: { overviewModel: runsOverviewModel }, modalModel }) => {
+    runsOverviewModel.pagination.provideDefaultItemsPerPage(estimateDisplayableRowsCount(
         TABLEROW_HEIGHT,
         PAGE_USED_HEIGHT,
     ));
 
-    return [
+    return h('', [
         h('.flex-row.header-container.pv2', [
-            filtersPanelPopover(model.runs, runsActiveColumns),
-            exportRunsButton(model),
+            filtersPanelPopover(runsOverviewModel, runsActiveColumns),
+            exportRunsButton({ runsOverviewModel, modalModel }),
         ]),
         h('.flex-column.w-100', [
-            table(runs, runsActiveColumns),
-            paginationComponent(model.runs.pagination),
+            table(runsOverviewModel.runs, runsActiveColumns),
+            paginationComponent(runsOverviewModel.pagination),
         ]),
-    ];
+    ]);
 };
 
 /**
@@ -59,10 +51,7 @@ const showRunsTable = (model, runs) => {
  * @param {Model} model global model to access functions
  * @returns {vnode} with button
  */
-const exportRunsButton = (model) =>
-    h('button.btn.btn-primary.w-15.h2.mlauto#export-runs-trigger', {
-        disabled: model.runs.runs.isSuccess() && model.runs.runs.payload.length === 0,
-        onclick: () => model.modal({ content: (modalModel) => exportRunsModal(model, modalModel), size: 'medium' }),
-    }, 'Export Runs');
-
-export default (model) => runsOverview(model);
+const exportRunsButton = ({ runsOverviewModel, modalModel }) => h('button.btn.btn-primary.w-15.h2.mlauto#export-runs-trigger', {
+    disabled: (runsOverviewModel.runs?.payload?.length ?? 0) === 0,
+    onclick: () => modalModel.display({ content: (modalModel) => exportRunsModal(runsOverviewModel, modalModel), size: 'medium' }),
+}, 'Export Runs');
