@@ -13,26 +13,25 @@
 
 import { h } from '/js/src/index.js';
 import { runsActiveColumns } from '../ActiveColumns/runsActiveColumns.js';
-import pick from '../../../utilities/pick.js';
 
 /**
  * Export form component, containing the fields to export, the export type and the export button
  *
- * @param {Model} model the global model
+ * @param {RunsOverviewModel} runsOverviewModel the runsOverviewModel
  * @param {array} runs the runs to export
  * @param {ModalHandler} modalHandler The modal handler, used to dismiss modal after export
  *
  * @return {vnode[]} the form component
  */
-const exportForm = (model, runs, modalHandler) => {
+const exportForm = (runsOverviewModel, runs, modalHandler) => {
     const exportTypes = ['JSON', 'CSV'];
-    const selectedRunsFields = model.runs.getSelectedRunsFields() || [];
-    const selectedExportType = model.runs.getSelectedExportType() || exportTypes[0];
+    const selectedRunsFields = runsOverviewModel.getSelectedRunsFields() || [];
+    const selectedExportType = runsOverviewModel.getSelectedExportType() || exportTypes[0];
     const runsFields = Object.keys(runsActiveColumns);
-    const enabled = runs && selectedRunsFields.length > 0 && Boolean(selectedExportType);
+    const enabled = selectedRunsFields.length > 0;
 
     return [
-        model.runs.isAllRunsTruncated
+        runsOverviewModel.isAllRunsTruncated
             ? h(
                 '#truncated-export-warning.warning',
                 `The runs export is limited to ${runs.length} entries, only the last runs will be exported (sorted by run number)`,
@@ -47,7 +46,7 @@ const exportForm = (model, runs, modalHandler) => {
         h('select#fields.form-control', {
             style: 'min-height: 20rem;',
             multiple: true,
-            onchange: ({ target }) => model.runs.setSelectedRunsFields(target.selectedOptions),
+            onchange: ({ target }) => runsOverviewModel.setSelectedRunsFields(target.selectedOptions),
         }, [
             ...runsFields
                 .filter((name) => !['id', 'actions'].includes(name))
@@ -67,7 +66,7 @@ const exportForm = (model, runs, modalHandler) => {
                     value: exportType,
                     checked: selectedExportType.length ? selectedExportType.includes(exportType) : false,
                     name: 'runs-export-type',
-                    onclick: () => model.runs.setSelectedExportType(exportType),
+                    onclick: () => runsOverviewModel.setSelectedExportType(exportType),
                 }),
                 h('label.form-check-label', {
                     for: id,
@@ -77,16 +76,10 @@ const exportForm = (model, runs, modalHandler) => {
         h('button.shadow-level1.btn.btn-success.mt2#send', {
             disabled: !enabled,
             onclick: async () => {
-                await model.runs.createRunsExport(
-                    runs.map((selectedRun) => {
-                        const entries = Object.entries(pick(selectedRun, selectedRunsFields));
-                        const formattedEntries = entries.map(([key, value]) => {
-                            const formatExport = runsActiveColumns[key].exportFormat || ((identity) => identity);
-                            return [key, formatExport(value, selectedRun)];
-                        });
-                        return Object.fromEntries(formattedEntries);
-                    }),
+                await runsOverviewModel.createRunsExport(
+                    runs,
                     'runs',
+                    runsActiveColumns,
                 );
                 modalHandler.dismiss();
             },
@@ -99,19 +92,19 @@ const errorDisplay = () => h('.danger', 'Data fetching failed');
 
 /**
  * A function to construct the exports runs screen
- * @param {Model} model Pass the model to access the defined functions
+ * @param {Model} runsOverviewModel Pass the model to access the defined functions
  * @param {ModalHandler} modalHandler The modal handler, used to dismiss modal after export
  * @return {vnode} Return the view of the inputs
  */
-export const exportRunsModal = (model, modalHandler) => {
-    const runsRemoteData = model.runs.allRuns;
+export const exportRunsModal = (runsOverviewModel, modalHandler) => {
+    const runsRemoteData = runsOverviewModel.allRuns;
 
     return h('div#export-runs-modal', [
         h('h2', 'Export Runs'),
         runsRemoteData.match({
             NotAsked: () => errorDisplay(),
-            Loading: () => exportForm(model, null, modalHandler),
-            Success: (payload) => exportForm(model, payload, modalHandler),
+            Loading: () => exportForm(runsOverviewModel, null, modalHandler),
+            Success: (payload) => exportForm(runsOverviewModel, payload, modalHandler),
             Failure: () => errorDisplay(),
         }),
     ]);
