@@ -25,6 +25,9 @@ import { tagOccurrencesBarChartComponent } from './charts/tagOccurrencesBarChart
 import { timeRangeFilter } from '../../components/Filters/common/filters/timeRangeFilter.js';
 import { eorReasonOccurrencesBarChartComponent } from './charts/eorReasonOccurrencesBarChartComponent.js';
 import { formatTimeRange } from '../../components/common/formatting/formatTimeRange.js';
+import spinner from '../../components/common/spinner.js';
+import errorAlert from '../../components/common/errorAlert.js';
+import { centeredComponent } from '../../components/common/centeredComponent.js';
 
 /**
  * Render the statistics page
@@ -61,7 +64,11 @@ export const StatisticsPage = ({ statisticsModel }) => {
                         h('h3', `Efficiency - ${periodLabel}`),
                         h(
                             '.flex-grow.chart-box',
-                            efficiencyChartComponent(fillStatistics, () => statisticsModel.notify()),
+                            efficiencyChartComponent(
+                                fillStatistics,
+                                () => statisticsModel.notify(),
+                                statisticsModel.timeRangeFilterModel.periodLabel,
+                            ),
                         ),
                     ],
                 }),
@@ -95,36 +102,50 @@ export const StatisticsPage = ({ statisticsModel }) => {
                 [STATISTICS_PANELS_KEYS.EFFICIENCY_PER_DETECTOR]: (panelModel) => {
                     const detectorsColors = ChartDarkColors;
                     return remoteDataDisplay(panelModel.data, {
-                        Success: (efficiencyPerDetectors) => [
-                            h('h3', `Efficiency per detector - ${periodLabel}`),
-                            h('.flex-row.g2', panelModel.detectors.map((detector, i) => switchInput(
-                                panelModel.getDetectorVisibility(detector),
-                                (visibility) => panelModel.setDetectorVisibility(detector, visibility),
-                                { key: detector, labelAfter: detector, color: detectorsColors[i] },
-                            ))),
-                            h('.flex-row.flex-grow.g5', [
-                                h(
-                                    '.flex-grow.chart-box',
-                                    detectorsEfficienciesComponent(
-                                        efficiencyPerDetectors,
-                                        false,
-                                        panelModel.activeDetectors,
-                                        detectorsColors.filter((_, i) => panelModel.getDetectorVisibility(panelModel.detectors[i])),
-                                        () => statisticsModel.notify(),
-                                    ),
-                                ),
-                                h(
-                                    '.flex-grow.chart-box',
-                                    detectorsEfficienciesComponent(
-                                        efficiencyPerDetectors,
-                                        true,
-                                        panelModel.activeDetectors,
-                                        detectorsColors.filter((_, i) => panelModel.getDetectorVisibility(panelModel.detectors[i])),
-                                        () => statisticsModel.notify(),
-                                    ),
-                                ),
-                            ]),
-                        ],
+                        Success: (efficiencyPerDetectors) => panelModel.detectors.match({
+                            NotAsked: () => centeredComponent('No detectors available'),
+                            Loading: () => spinner(),
+                            Success: (detectors) => {
+                                const activeDetectors = [];
+                                for (const detector of detectors) {
+                                    if (panelModel.getDetectorVisibility(detector)) {
+                                        activeDetectors.push(detector);
+                                    }
+                                }
+
+                                return [
+                                    h('h3', `Efficiency per detector - ${periodLabel}`),
+                                    h('.flex-row.g2', detectors.map((detector, i) => switchInput(
+                                        panelModel.getDetectorVisibility(detector),
+                                        (visibility) => panelModel.setDetectorVisibility(detector, visibility),
+                                        { key: detector, labelAfter: detector, color: detectorsColors[i] },
+                                    ))),
+                                    h('.flex-row.flex-grow.g5', [
+                                        h(
+                                            '.flex-grow.chart-box',
+                                            detectorsEfficienciesComponent(
+                                                efficiencyPerDetectors,
+                                                false,
+                                                activeDetectors,
+                                                detectorsColors.filter((_, i) => panelModel.getDetectorVisibility(detectors[i])),
+                                                () => statisticsModel.notify(),
+                                            ),
+                                        ),
+                                        h(
+                                            '.flex-grow.chart-box',
+                                            detectorsEfficienciesComponent(
+                                                efficiencyPerDetectors,
+                                                true,
+                                                activeDetectors,
+                                                detectorsColors.filter((_, i) => panelModel.getDetectorVisibility(detectors[i])),
+                                                () => statisticsModel.notify(),
+                                            ),
+                                        ),
+                                    ]),
+                                ];
+                            },
+                            Failure: (error) => centeredComponent(errorAlert(error)),
+                        }),
                     });
                 },
                 [STATISTICS_PANELS_KEYS.LOG_TAG_OCCURRENCES]: (remoteData) => remoteDataDisplay(remoteData, {
