@@ -10,10 +10,10 @@
  *  granted to it by virtue of its status as an Intergovernmental Organization
  *  or submit itself to any jurisdiction.
  */
-import { Observable } from '/js/src/index.js';
+import { Observable, RemoteData } from '/js/src/index.js';
 import { TabbedPanelModel } from '../../components/TabbedPanel/TabbedPanelModel.js';
 import { buildUrl } from '../../utilities/fetch/buildUrl.js';
-import { RemoteDataFetcher } from '../../utilities/fetch/RemoteDataFetcher.js';
+import { RemoteDataSource } from '../../utilities/fetch/RemoteDataSource.js';
 import { TimeRangeFilterModel } from '../../components/Filters/common/filters/TimeRangeFilterModel.js';
 import { getStartOfDay, getStartOfYear, MILLISECONDS_IN_ONE_DAY } from '../../utilities/dateUtils.js';
 
@@ -52,6 +52,8 @@ export class StatisticsPageModel extends Observable {
                 this._tabbedPanelModel.period = this._timeRangeFilterModel.normalized;
             }
         });
+        this._timeRangeFilterModel.bubbleTo(this);
+
         // Notify to update panel period
         this._timeRangeFilterModel.notify();
     }
@@ -96,8 +98,8 @@ class StatisticsTabbedPanelModel extends TabbedPanelModel {
      */
     constructor() {
         super(Object.values(STATISTICS_PANELS_KEYS));
-        this._remoteDataFetcher = new RemoteDataFetcher();
-        this._remoteDataFetcher.bubbleTo(this);
+        this._remoteDataFetcher = new RemoteDataSource();
+        this._remoteDataFetcher.observableData.bubbleTo(this);
         this._detectorEfficiencyTabModel = new DetectorsEfficienciesTabModel();
         this._detectorEfficiencyTabModel.bubbleTo(this);
     }
@@ -177,7 +179,7 @@ class StatisticsTabbedPanelModel extends TabbedPanelModel {
             case STATISTICS_PANELS_KEYS.TIME_BETWEEN_RUNS_DISTRIBUTION:
             case STATISTICS_PANELS_KEYS.LOG_TAG_OCCURRENCES:
             case STATISTICS_PANELS_KEYS.EOR_REASON_OCCURRENCES:
-                return this._remoteDataFetcher.data;
+                return this._remoteDataFetcher.observableData.current;
             case STATISTICS_PANELS_KEYS.EFFICIENCY_PER_DETECTOR:
                 return this._detectorEfficiencyTabModel;
         }
@@ -207,14 +209,32 @@ class DetectorsEfficienciesTabModel extends Observable {
         super();
 
         this._period = null;
-        this._remoteDataFetcher = new RemoteDataFetcher();
-        this._remoteDataFetcher.bubbleTo(this);
 
         /**
-         * @type {string[]}
+         * @type {RemoteData<string[]>}
          * @private
          */
-        this._detectors = [];
+        this._detectors = RemoteData.notAsked();
+
+        this._remoteDataFetcher = new RemoteDataSource();
+        this._remoteDataFetcher.observableData.observe(() => {
+            this._detectors = this._remoteDataFetcher.observableData.current.match({
+                NotAsked: () => RemoteData.notAsked(),
+                Loading: () => RemoteData.loading(),
+                Success: (detectorEfficiency) => RemoteData.success(detectorEfficiency.reduce((detectors, { efficiencies }) => {
+                    for (const detector in efficiencies) {
+                        if (!detectors.includes(detector)) {
+                            detectors.push(detector);
+                        }
+                    }
+                    detectors.sort((a, b) => a.localeCompare(b));
+                    return detectors;
+                }, [])),
+                Failure: (error) => RemoteData.failure(error),
+            });
+
+            this.notify();
+        });
 
         /**
          * @type {Map<string, boolean>}
@@ -231,17 +251,7 @@ class DetectorsEfficienciesTabModel extends Observable {
         this._remoteDataFetcher.fetch(buildUrl(
             '/api/statistics/runs/detectorEfficiency',
             { from: this._period.from, to: this._period.to },
-        )).then((detectorEfficiency) => {
-            this._detectors = detectorEfficiency.reduce((detectors, { efficiencies }) => {
-                for (const detector in efficiencies) {
-                    if (!detectors.includes(detector)) {
-                        detectors.push(detector);
-                    }
-                }
-                detectors.sort((a, b) => a.localeCompare(b));
-                return detectors;
-            }, []);
-        });
+        ));
     }
 
     /**
@@ -250,7 +260,7 @@ class DetectorsEfficienciesTabModel extends Observable {
      * @return {RemoteData} the tab data
      */
     get data() {
-        return this._remoteDataFetcher.data;
+        return this._remoteDataFetcher.observableData.current;
     }
 
     /**
@@ -274,7 +284,7 @@ class DetectorsEfficienciesTabModel extends Observable {
     /**
      * Return the list of all the detectors that have statistics
      *
-     * @return {string[]} the detectors
+     * @return {RemoteData<string[]>} the detectors
      */
     get detectors() {
         return this._detectors;
@@ -300,20 +310,5 @@ class DetectorsEfficienciesTabModel extends Observable {
     setDetectorVisibility(detector, visibility) {
         this._detectorsVisibility.set(detector, visibility);
         this.notify();
-    }
-
-    /**
-     * Return the list of detector for which efficiency should be displayed
-     *
-     * @return {string[]} the list of detectors
-     */
-    get activeDetectors() {
-        const ret = [];
-        for (const detector of this._detectors) {
-            if (this.getDetectorVisibility(detector)) {
-                ret.push(detector);
-            }
-        }
-        return ret;
     }
 }
