@@ -15,6 +15,7 @@ import { lineChartComponent } from '../../../components/common/chart/lineChart/l
 import { h } from '/js/src/index.js';
 import { formatPercentage } from '../../../utilities/formatting/formatPercentage.js';
 import { ChartColors } from '../chartColors.js';
+import { centeredComponent } from '../../../components/common/centeredComponent.js';
 
 /**
  * Return the efficiency chart component for a given fill statistics
@@ -28,7 +29,7 @@ import { ChartColors } from '../chartColors.js';
  */
 export const detectorsEfficienciesComponent = (detectorEfficiencies, net, detectors, colors, onPointHover) => {
     if (detectors.length === 0) {
-        return h('.w-100.h-100.flex-row.items-center.justify-center', 'No detector selected');
+        return centeredComponent('No detector selected');
     }
 
     return lineChartComponent(
