@@ -17,24 +17,36 @@ import { table } from '../../../components/common/table/table.js';
 import { lhcPeriodsActiveColumns } from '../ActiveColumns/lhcPeriodsActiveColumns.js';
 import { paginationComponent } from '../../../components/Pagination/paginationComponent.js';
 import { filtersPanelPopover } from '../../../components/Filters/common/filtersPanelPopover.js';
+import { estimateDisplayableRowsCount } from '../../../utilities/estimateDisplayableRowsCount.js';
+
+const TABLEROW_HEIGHT = 35;
+// Estimate of the navbar and pagination elements height total; Needs to be updated in case of changes;
+const PAGE_USED_HEIGHT = 215;
 
 /**
  * Render LHC Periods overview model
  * @param {Model} model The overall model object.
  * @returns {Component} The overview screen
  */
-export const LhcPeriodsOverviewPage = ({ lhcPeriods: { overviewModel: lhcPeriodsOverviewModel } }) => h('', {
-    onremove: () => lhcPeriodsOverviewModel.reset(),
-}, [
-    h('.flex-row.flex-grow.items-start', filtersPanelPopover(lhcPeriodsOverviewModel, lhcPeriodsActiveColumns)),
-    h('.w-100.flex-column', [
-        table(
-            lhcPeriodsOverviewModel.lhcPeriods,
-            lhcPeriodsActiveColumns,
-            { classes: '.table-sm' },
-            null,
-            { sort: lhcPeriodsOverviewModel.overviewSortModel },
-        ),
-        paginationComponent(lhcPeriodsOverviewModel.pagination),
-    ]),
-]);
+export const LhcPeriodsOverviewPage = ({ lhcPeriods: { overviewModel: lhcPeriodsOverviewModel } }) => {
+    lhcPeriodsOverviewModel.pagination.provideDefaultItemsPerPage(estimateDisplayableRowsCount(
+        TABLEROW_HEIGHT,
+        PAGE_USED_HEIGHT,
+    ));
+
+    return h('', {
+        onremove: () => lhcPeriodsOverviewModel.reset(),
+    }, [
+        h('.flex-row.flex-grow.items-start', filtersPanelPopover(lhcPeriodsOverviewModel, lhcPeriodsActiveColumns)),
+        h('.w-100.flex-column', [
+            table(
+                lhcPeriodsOverviewModel.lhcPeriods,
+                lhcPeriodsActiveColumns,
+                { classes: '.table-sm' },
+                null,
+                { sort: lhcPeriodsOverviewModel.overviewSortModel },
+            ),
+            paginationComponent(lhcPeriodsOverviewModel.pagination),
+        ]),
+    ]);
+};
