@@ -47,7 +47,13 @@ class GRPCRunController {
 
         return this.runToGRPC(await this.runService.update(
             { runNumber },
-            { runPatch: run, relations: { runTypeName: relations.runType?.name } },
+            {
+                runPatch: run,
+                relations: {
+                    runTypeName: relations.runType?.name,
+                    eorReasons: relations.eorReasons,
+                },
+            },
         ));
     }
 
@@ -87,6 +93,11 @@ class GRPCRunController {
         if (gRPCRunRequest.runType) {
             relations.runType = { name: gRPCRunRequest.runType };
             delete gRPCRunRequest.runType;
+        }
+
+        if (gRPCRunRequest.eorReasons) {
+            relations.eorReasons = gRPCRunRequest.eorReasons;
+            delete gRPCRunRequest.eorReasons;
         }
 
         const { lhcPeriod } = gRPCRunRequest;
