@@ -87,6 +87,7 @@ class LogService {
 
     /**
      * Return all the logs that correspond to a given list of ids with a default list of shallow relations
+     * and also adds the number of replies to each log.
      *
      * @param {number[]} logIds the list of ids of logs to fetch
      * @return {Promise<Log[]>} the resulting logs
@@ -98,7 +99,9 @@ class LogService {
             .include({ association: 'lhcFills', attributes: ['fillNumber'] })
             .include({ association: 'environments', attributes: ['id'] });
 
-        return (await LogRepository.findAll(logQueryBuilder)).map(logAdapter.toEntity);
+        const logs = await LogRepository.findAll(logQueryBuilder);
+        const logEntities = logs.map(logAdapter.toEntity);
+        return LogRepository.addChildrenCountByLog(logEntities);
     }
 
     /**
