@@ -276,13 +276,30 @@ class RunService {
  * @returns {Promise<undefined|Error>} - promise on result of db queries
  */
 const updateEorReasonsOnRun = async (runId, userName, eorReasonPatches) => {
-    let reasonTypes = await ReasonTypeRepository.findAll();
-    reasonTypes = reasonTypes.map((reason) => reason.id);
+    const reasonTypes = await ReasonTypeRepository.findAll();
 
-    const allReceivedReasonsExists = eorReasonPatches.every((eorReasonPatch) => reasonTypes.includes(eorReasonPatch.reasonTypeId));
+    /**
+     * @type {Map<number|string, number>}
+     */
+    const reasonTypesMap = new Map(reasonTypes
+        .map(({ id, category, title }) => [
+            // Set in the same map possibility to index by reasontypeId or by title/category
+            [id, id],
+            [JSON.stringify([category, title]), id],
+        ])
+        // Convert the array of two values into 2 values in the main array
+        .flat());
 
-    if (!allReceivedReasonsExists) {
-        throw new Error('Provided reason types do not exist');
+    const validPatches = [];
+    for (const patch of eorReasonPatches) {
+        // Use reasonTypeId if it exists, otherwise use category+title if it exists to get the reasonTypeId
+        const reasonTypeId = reasonTypesMap.get(patch.reasonTypeId ?? JSON.stringify([patch.category, patch.title])) ?? null;
+        // Add to the validPatches array if valid, if the patch is invalid ignore it
+        if (reasonTypeId !== null) {
+            // Set the reasonTypeId for further processing
+            patch.reasonTypeId = reasonTypeId;
+            validPatches.push(patch);
+        }
     }
 
     const toKeepEorReasonsIds = []; // EorReasons with an ID already, means exist in DB;
@@ -290,7 +307,7 @@ const updateEorReasonsOnRun = async (runId, userName, eorReasonPatches) => {
      * @type {Partial<SequelizeEorReason>[]}
      */
     const newEorReasons = []; // EorReasons with no ID, need to be added in DB;
-    eorReasonPatches.forEach(({ id, reasonTypeId, description }) => {
+    validPatches.forEach(({ id, reasonTypeId, description }) => {
         if (id) {
             toKeepEorReasonsIds.push(id);
         } else {
