@@ -27,11 +27,22 @@ const RUN_QUALITY_CHANGE_TAGS = ['DPG', 'RC'];
  * @param {Object} transaction the current transaction
  * @param {User|null} user if not null, the user mentioned in the log as the author of the update
  * @param {string} changeReason the reason of the run quality change
+ * @param {string} lhcPeriodName the LHC period of the run
  * @return {Promise<void>} resolve when log has been created
  * @private
  */
-exports.logRunQualityChange = async (runNumber, previousQuality, newQuality, transaction, user, changeReason) => {
-    const textParts = [`The run quality for run ${runNumber} has been changed from ${previousQuality} to ${newQuality}`];
+exports.logRunQualityChange = async (
+    runNumber,
+    previousQuality,
+    newQuality,
+    transaction,
+    user,
+    changeReason,
+    lhcPeriodName,
+) => {
+    const textParts =
+        [`The run quality for run ${runNumber} from period ${lhcPeriodName} has been changed from ${previousQuality} to ${newQuality}`];
+
     if (user) {
         textParts.push(`by ${user.name}`);
     }
