@@ -245,7 +245,7 @@ class GetAllLogsUseCase {
             return LogRepository.findAndCountAll(queryBuilder);
         });
 
-        const rowsWithReplies = await LogRepository.addChildrenCountByRootLog(rows);
+        const rowsWithReplies = await LogRepository.addChildrenCountByLog(rows);
 
         return {
             count,
