@@ -45,7 +45,7 @@ class GetLogsByRunUseCase {
         });
 
         if (logs) {
-            const logsWithReplies = await LogRepository.addChildrenCountByRootLog(logs);
+            const logsWithReplies = await LogRepository.addChildrenCountByLog(logs);
             return logsWithReplies.map(logAdapter.toEntity);
         } else {
             return null;
