@@ -63,7 +63,7 @@ class GetLogsByTagUseCase {
                 },
             });
 
-            const logsWithCount = await LogRepository.addChildrenCountByRootLog(logs);
+            const logsWithCount = await LogRepository.addChildrenCountByLog(logs);
 
             return { logs: logsWithCount, count };
         });
